requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
cssselect2==0.7.0
distlib==0.3.5
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-pwa==1.1.0
filelock==3.8.0
frozenlist==1.3.3
geographiclib==2.0
html5lib==1.1
idna==3.4
incremental==22.10.0
kombu==5.2.4
lxml==4.9.2
msgpack==1.0.4
multidict==6.0.4
numpy==1.25.0
oscrypto==1.3.0
pandas==2.0.2
paystack==1.5.0
Pillow==9.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.38
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pypdf==3.11.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytube==12.1.0
pytz==2023.3
pywebcopy==7.0.2
PyYAML==6.0
qrcode==7.4.2
redis==4.5.4
reportlab==3.6.13
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
twilio==8.2.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.11
yarl==1.9.2
zope.interface==6.0
